requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn